use own fromEntries
